README.md
This is a very Simple Accounting System for newbies in C#. This project comprises of Chart of Accounts, Simple transactions, a System for Debit And Credit of customers, and other income or expense transactions. It generates reports which are Profit/Loss Statement, Customer Statements, and cashbook statements. This C# project was developed with SQL Server as the back end and Crystal Report. You will certainly learn a lot from the code behind it.

The system works as the following:
System users must populate first the list of the Customers, Cash/Banks, and Chart of Accounts.
Then Encode the encode/store the customer's transaction and post it as a Debit or Credit.
After that, the Profit/Loss Report will be automatically generated and the other reports will be based on the filter options such as the customer name and date range in the Customer Report Statement.
Features
Manage Customer List
Manage Cash/Bank List
Manage Chart of Accounts
Manage Cash/Bank Transactions
Generates Reports
Profit/Loss Report
Customer Statement of Account Report
Cash/Bank Statement Report